move read_streams to output_processor
